+++ refena/lib/src/notifier/types/pure_notifier.dart
@@ -7,5 +7,5 @@ import 'package:refena/src/ref.dart';
 /// It is used in combination with dependency injection,
 /// where you provide the dependencies via constructor.
 abstract class PureNotifier<T> extends BaseSyncNotifier<T> {
-  PureNotifier({String? debugLabel}) : super(debugLabel: debugLabel);
+  PureNotifier();
 }